+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## [Unreleased] 
 
+### Added
+
 - [TD-2951] `profile_structure` permission
 
 ### Fixed
 
 - [TD-3235] Fallback of uncontroller responses on metadata controller
+
+### Removed
+
+- [TD-3421] remove `/data_structures/search/source_alias` endpoint
 
 ## [4.15.0] 2021-03-08